[Elliptic] Enforce continuity of the multigrid precon. output.
(Of course, this is only for continuous discretizations.)

This can make a difference for some of the EXTRAP initial guess methods
that are on the verge of instability.

Since this doesn't seem to matter in most cases, it would be better to
enforce the continuity only at the the output of the linear solver, but
there's an encapsulation problem:  the linear solver object does not
know about discretizations.